Luo uusi abstraktiotaso tiedostomanagereille: VaiheTiedostoManager extends TiedostoManager. Tämä valmistelee tulevia lausuntopyyntöjä varten - niillä ei ole julkaisuja.
